refactor: change pages to app directory
