+++ src/app/Carteira/index.ts
@@ -1,0 +1,2 @@
+export * from './hook'
+export * from './page'
