--- src/app/Wallet/index.ts
@@ -1,2 +1,0 @@
-export * from './actions'
-export * from './page'
