+++ src/components/index.ts
@@ -1,3 +1,10 @@
+export * from './ErrorPage'
+export * from '../app/Home/page'
+export * from './Loading'
+export * from './Modal'
+export * from './Button'
+export * from './Select'
+
 export * from "./ErrorPage";
 export * from "./Home";
 export * from "./LandingPage";
